Addressing numeric conversion warnings in Mass.
#preflight 63764fbc232e3d12cba18ccd
#review-23156037
#jira UE-166285

[CL 23175896 by Stephen Holmes in ue5-main branch]

+++ Engine/Plugins/AI/MassAI/Source/MassAIBehavior/Private/Evaluators/MassComponentHitEvaluator.cpp
@@ -32,9 +32,9 @@ void FMassComponentHitEvaluator::Tick(FStateTreeExecutionContext &Context, const
 		// @todo: This is a bit of a kludge to expose an event to StateTree.
 		const UWorld* World = Context.GetWorld();
 		check(World);
-		const float CurrentTime = World->GetTimeSeconds();
-		const float TimeSinceHit = CurrentTime - HitResult->HitTime;
-		constexpr float HitEventDuration = 0.1f;
+		const double CurrentTime = World->GetTimeSeconds();
+		const double TimeSinceHit = CurrentTime - HitResult->HitTime;
+		constexpr double HitEventDuration = 0.1;
 		if (TimeSinceHit < HitEventDuration)
 		{
 			MASSBEHAVIOR_LOG(VeryVerbose, TEXT("Got hit"));
