+++ Engine/Plugins/Runtime/MassEntity/Source/MassEntityTestSuite/Private/MassCommandsTest.cpp
@@ -31,8 +31,8 @@ struct FCommands_FragmentInstanceList : FEntityTestBase
 
 		for (int i = 0; i < Count; ++i)
 		{
-			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(IntEntities[i], FTestFragment_Int(i), FTestFragment_Float(i));
-			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(FloatEntities[i], FTestFragment_Int(i), FTestFragment_Float(i));
+			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(IntEntities[i], FTestFragment_Int(i), FTestFragment_Float((float)i));
+			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(FloatEntities[i], FTestFragment_Int(i), FTestFragment_Float((float)i));
 		}
 
 		EntityManager->FlushCommands();
@@ -73,7 +73,7 @@ struct FCommands_FragmentMemoryCleanup : FEntityTestBase
 		const FMassArchetypeHandle ArrayArchetype = EntityManager->CreateArchetype(MakeArrayView(ArrayFragmentTypes, 1));
 		const FMassArchetypeHandle ArrayIntArchetype = EntityManager->CreateArchetype(MakeArrayView(ArrayFragmentTypes, 2));
 		const int32 EntitiesPerChunk = EntityManager->DebugGetArchetypeEntitiesCountPerChunk(ArrayArchetype);
-		const int32 Count = EntitiesPerChunk * 2.5f;
+		const int32 Count = (int32)(EntitiesPerChunk * 2.5f);
 		
 		TArray<FMassEntityHandle> Entities;
 		EntityManager->BatchCreateEntities(ArrayArchetype, Count, Entities);
@@ -113,13 +113,13 @@ struct FCommands_BuildEntitiesWithFragments : FEntityTestBase
 	virtual bool InstantTest() override
 	{
 		const int32 EntitiesPerChunk = EntityManager->DebugGetArchetypeEntitiesCountPerChunk(FloatsIntsArchetype);
-		const int32 Count = EntitiesPerChunk * 2.5f;
+		const int32 Count = (int32)(EntitiesPerChunk * 2.5f);
 
 		TArray<FMassEntityHandle> Entities;
 		for (int i = 0; i < Count; ++i)
 		{
 			Entities.Add(EntityManager->ReserveEntity());
-			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(Entities.Last(), FTestFragment_Int(i), FTestFragment_Float(i));
+			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(Entities.Last(), FTestFragment_Int(i), FTestFragment_Float((float)i));
 		}
 
 		AITEST_EQUAL(TEXT("All entities created should be in ArrayArchetype"), EntityManager->DebugGetArchetypeEntitiesCount(FloatsIntsArchetype), 0);
@@ -155,7 +155,7 @@ struct FCommands_BuildEntitiesInHoles : FEntityTestBase
 		for (int i = 0; i < EntitiesPerChunk; ++i)
 		{
 			Entities.Add(EntityManager->ReserveEntity());
-			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(Entities.Last(), FTestFragment_Int(i), FTestFragment_Float(i));
+			EntityManager->Defer().PushCommand<FMassCommandAddFragmentInstances>(Entities.Last(), FTestFragment_Int(i), FTestFragment_Float((float)i));
 		}
 
 		AITEST_EQUAL(TEXT("All entities created should be in ArrayArchetype"), EntityManager->DebugGetArchetypeEntitiesCount(FloatsIntsArchetype), Count / 2);
@@ -179,13 +179,13 @@ struct FCommands_BuildEntitiesWithFragmentInstances : FEntityTestBase
 	virtual bool InstantTest() override
 	{
 		const int32 EntitiesPerChunk = EntityManager->DebugGetArchetypeEntitiesCountPerChunk(FloatsIntsArchetype);
-		const int32 Count = EntitiesPerChunk * 2.5f;
+		const int32 Count = (int32)(EntitiesPerChunk * 2.5f);
 
 		TArray<FMassEntityHandle> Entities;
 		for (int i = 0; i < Count; ++i)
 		{
 			Entities.Add(EntityManager->ReserveEntity());
-			EntityManager->Defer().PushCommand<FMassCommandBuildEntity>(Entities.Last(), FTestFragment_Int(i), FTestFragment_Float(i));
+			EntityManager->Defer().PushCommand<FMassCommandBuildEntity>(Entities.Last(), FTestFragment_Int(i), FTestFragment_Float((float)i));
 		}
 
 		AITEST_EQUAL(TEXT("All entities created should be in ArrayArchetype"), EntityManager->DebugGetArchetypeEntitiesCount(FloatsIntsArchetype), 0);
